+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The following is a curated list of changes in the Enact limestone module, newest
 ### Changed
 
 - All components in `light` skin and high-contrast mode styling to match the latest GUI
+- `limestone/Card` to display `captions` on focus
 - `limestone/CheckboxItem` styling to match the latest GUI
 - `limestone/FlexiblePopupPanels` styling to match the latest GUI
 - `limestone/FormCheckboxItem` styling to match the latest GUI
@@ -25,6 +26,7 @@ The following is a curated list of changes in the Enact limestone module, newest
 ### Fixed
 
 - `limestone/Card` for `focusRing` mode
+- `limestone/Card` `background` for captions
 - `limestone/Icon` medium size
 - `limestone/Popup` to restore focus properly when popup hides
 - `limestone/QuickGuidePanels` navigation buttons position